+++ joulescope_ui/widgets/statistics_record/statistics_record.py
@@ -12,24 +12,11 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from pyjls import Writer, SignalType, DataType
 from joulescope_ui import pubsub_singleton, register, CAPABILITIES, Metadata, get_topic_name
 from .statistics_record_config_widget import StatisticsRecordConfigDialog
-import copy
+from pyjoulescope_driver import time64
 import logging
 import numpy as np
-
-
-class ChunkMeta:
-    NOTES = 0
-    UI_META = 0x801
-
-
-_DTYPE_MAP = {
-    'f32': DataType.F32,
-    'u8': DataType.U8,
-    'u1': DataType.U1,
-}
 
 
 @register
@@ -41,8 +28,14 @@ class StatisticsRecord:
         '!stop': Metadata('bool', 'Recording stopped', flags=['ro', 'skip_undo']),
     }
 
-    def __init__(self, topic, filename):
+    def __init__(self, topic, filename, config=None):
         parent = pubsub_singleton.query('registry/app/instance')
+        time_format = config.get('time_format', None)
+        if time_format is None:
+            time_format = 'relative'
+        if time_format not in ['relative', 'UTC', 'off']:
+            raise ValueError(f'invalid time_format {time_format}')
+        self._time_format = time_format
         self._topic = topic
         self._log = logging.getLogger(f'{__name__}.obj')
         self.CAPABILITIES = [CAPABILITIES.STATISTIC_STREAM_SINK]
@@ -50,6 +43,7 @@ class StatisticsRecord:
         self._on_data_fn = self._on_data
         self._file = open(filename, 'wt')
         self._offsets = None
+        self._time_format_str = ''
 
         pubsub_singleton.publish('registry/paths/actions/!mru_save', filename)
         pubsub_singleton.register(self, parent=parent)
@@ -58,12 +52,7 @@ class StatisticsRecord:
     def _on_data(self, topic, value):
         if self._file is None:
             return
-        dt = value['time']['delta']['value']
-        freq = 1.0 / dt
-        freq_log10 = int(np.ceil(max(0, np.log10(freq))))
-        time_format = f'%.{freq_log10}f'
-        hdr = '#time,current,voltage,power,charge,energy\n'
-        t = value['time']['range']['value'][1]
+        t = value['time']['utc']['value'][1]  # time64 format
         i = value['signals']['current']['avg']['value']
         v = value['signals']['voltage']['avg']['value']
         p = value['signals']['power']['avg']['value']
@@ -76,12 +65,37 @@ class StatisticsRecord:
                 'charge': c,
                 'energy': e,
             }
+            hdr = '#' if self._time_format == 'off' else '#time,'
+            hdr += 'current,voltage,power,charge,energy\n'
             self._file.write(hdr)
-        t -= self._offsets['time']
+            freq = 1.0 / value['time']['delta']['value']
+            freq_log10 = int(np.ceil(max(0, np.log10(freq))))
+            if self._time_format == 'relative':
+                fmt_str = f'%.{freq_log10}f,'
+                def relative_format(t_now):
+                    dt = (t_now - self._offsets['time'])
+                    dt /= time64.SECOND
+                    return fmt_str % dt
+                self._time_format_fn = relative_format
+            elif self._time_format == 'UTC':
+                def utc_format(t_now):
+                    d = time64.as_datetime(t_now)
+                    if freq_log10 == 0:
+                        s = d.strftime('%Y%m%dT%H%M%SZ,')
+                    else:
+                        s = d.strftime('%Y%m%dT%H%M%S.%f')
+                        if freq_log10 < 6:
+                            s = s[:-6+freq_log10]
+                        s += 'Z,'
+                    return s
+                self._time_format_fn = utc_format
+            else:
+                self._time_format_fn = lambda t_now: ''
+
         c -= self._offsets['charge']
         e -= self._offsets['energy']
-        line = (time_format + ',%g,%g,%g,%g,%g\n') % (t, i, v, p, c, e)
-        self._file.write(line)
+        line = '%g,%g,%g,%g,%g\n' % (i, v, p, c, e)
+        self._file.write(self._time_format_fn(t) + line)
 
     def on_action_stop(self, value):
         self._log.info('stop')
@@ -94,9 +108,10 @@ class StatisticsRecord:
 
     @staticmethod
     def on_cls_action_start(pubsub, topic, value):
+        config = value.get('config', {})
         StatisticsRecord._log.info('on_cls_action_start')
-        for topic, filename in value:
-            obj = StatisticsRecord(topic, filename)
+        for topic, filename in value['sources']:
+            obj = StatisticsRecord(topic, filename, config)
             StatisticsRecord._instances.append(obj)
 
     @staticmethod